Split CircularSliderTool
